% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{gen_dynamic}
\alias{gen_dynamic}
\title{Generate dynamic network and simulate games}
\usage{
gen_dynamic(
  N,
  L,
  aver_k = 6,
  net_type = "BA",
  game_type = "PDG",
  visualize = FALSE
)
}
\arguments{
\item{N}{Number of nodes}

\item{L}{Number of iterations}

\item{aver_k}{Average degree}

\item{net_type}{Network type ("BA", "ER", "SW")}

\item{game_type}{Game type ("PDG", "SDG", "DG", "SUG", "CG", "SG", "EG", "SeqG", "RG", "ZZG", "NCG", "IIG")}

\item{visualize}{Logical, if TRUE, visualize the network}
}
\value{
A list containing the initial adjacency matrix, final adjacency matrix, game results, and strategies
}
\description{
Generate dynamic network and simulate games
}
