% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{generate_network}
\alias{generate_network}
\title{Generate network based on type}
\usage{
generate_network(N, aver_k, net_type)
}
\arguments{
\item{N}{Number of nodes}

\item{aver_k}{Average degree}

\item{net_type}{Network type ("BA", "ER", "SW")}
}
\value{
An igraph network
}
\description{
Generate network based on type
}
