% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{initialize_payoff_matrix}
\alias{initialize_payoff_matrix}
\title{Initialize payoff matrix based on game type}
\usage{
initialize_payoff_matrix(game_type)
}
\arguments{
\item{game_type}{Game type ("PDG", "SDG", "DG", "SUG", "CG", "SG", "EG", "SeqG", "RG", "ZZG", "NCG", "IIG")}
}
\value{
A payoff matrix
}
\description{
Initialize payoff matrix based on game type
}
