% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{initialize_strategies}
\alias{initialize_strategies}
\title{Initialize strategies based on game type}
\usage{
initialize_strategies(N, game_type)
}
\arguments{
\item{N}{Number of nodes}

\item{game_type}{Game type ("PDG", "SDG", "DG", "SUG", "CG", "SG", "EG", "SeqG", "RG", "ZZG", "NCG", "IIG")}
}
\value{
A vector of strategies
}
\description{
Initialize strategies based on game type
}
