% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_games}
\alias{simulate_games}
\title{Simulate games on the network}
\usage{
simulate_games(G, N, L, game_type, M, S)
}
\arguments{
\item{G}{An igraph network}

\item{N}{Number of nodes}

\item{L}{Number of iterations}

\item{game_type}{Game type}

\item{M}{Payoff matrix}

\item{S}{Strategies}
}
\value{
A list of results
}
\description{
Simulate games on the network
}
