% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate input parameters}
\usage{
validate_parameters(N, L, aver_k, net_type, game_type)
}
\arguments{
\item{N}{Number of nodes}

\item{L}{Number of iterations}

\item{aver_k}{Average degree}

\item{net_type}{Network type ("BA", "ER", "SW")}

\item{game_type}{Game type ("PDG", "SDG", "DG", "SUG", "CG", "SG", "EG", "SeqG", "RG", "ZZG", "NCG", "IIG")}
}
\description{
Validate input parameters
}
