% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize the network with strategies}
\usage{
visualize_network(G, S)
}
\arguments{
\item{G}{An igraph network}

\item{S}{Strategies}
}
\description{
Visualize the network with strategies
}
